refactor + more log
